+++ flutter_theoplayer_sdk/lib/theoplayer_state.dart
@@ -29,9 +29,22 @@ class PlayerState {
   List<TimeRange?> buffered = [];
   List<TimeRange?> seekable = [];
   List<TimeRange?> played = [];
+  bool allowBackgroundPlayback = false;
   String? error;
 
   bool isInitialized = false;
+
+  PresentationMode _presenationMode = PresentationMode.INLINE;
+
+  PresentationMode get presentationMode {
+    return _presenationMode;
+  }
+
+  set presentationMode(PresentationMode presentationMode) {
+    _presenationMode = presentationMode;
+    eventManager.dispatchEvent(PresentationModeChangeEvent(currentTime: currentTime, presentationMode: presentationMode));
+    _stateChangeListener?.call();
+  }
 
   PlayerState() {
     resetState();
@@ -224,6 +237,7 @@ class PlayerState {
     seekable = [];
     played = [];
     error = null;
+    _presenationMode = PresentationMode.INLINE;
   }
 
   /// Method to clean the internal state on player dispose.