+++ flutter_theoplayer_sdk_platform_interface/lib/theoplayer_platform_interface.dart
@@ -3,12 +3,14 @@ import 'package:theoplayer_platform_interface/theopalyer_config.dart';
 import 'package:theoplayer_platform_interface/theoplayer_view_controller_interface.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-typedef THEOplayerViewCreatedCallback = void Function(THEOplayerViewController controller);
+typedef THEOplayerViewCreatedCallback = void Function(THEOplayerViewController controller, BuildContext context);
+typedef InitializeNativeResultCallback = void Function(int playerId);
 
 abstract class TheoplayerPlatform extends PlatformInterface {
   /// Constructs a TheoplayerPlatform.
   TheoplayerPlatform() : super(token: _token);
 
+  static final int UNSUPPORTED_TEXTURE_ID = -1;
   static final Object _token = Object();
 
   static TheoplayerPlatform _instance = NotImplementedTHEOplayerPlatform();
@@ -26,7 +28,18 @@ abstract class TheoplayerPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Widget buildView(BuildContext context, THEOplayerConfig theoPlayerConfig, THEOplayerViewCreatedCallback createdCallback) {
+  /// Initializes the native player on the underlying platform
+  /// - for PlatformViews nothing happens here (yet)
+  /// - for Textures the player will be created
+  void initalize(THEOplayerConfig theoPlayerConfig, InitializeNativeResultCallback callback) {
+    throw UnimplementedError('initalize() has not been implemented.');
+  }
+
+  /// Builds the native you on the underlying platform
+  /// - for PlatormViews the player initialization happens here too
+  /// - for Textures the Texture gets created here based on the work done by the [initalize] call
+  Widget buildView(BuildContext context, THEOplayerConfig theoPlayerConfig, THEOplayerViewCreatedCallback createdCallback, int textureId) {
+    // TODO: maybe better to split this method for PlatformViews and Textures
     throw UnimplementedError('buildView(BuildContext) has not been implemented.');
   }
 }