+++ flutter_theoplayer_sdk_platform_interface/pigeons/pigeons_merged.dart
@@ -209,6 +209,10 @@ abstract class THEOplayerNativeAPI {
 
   List<TimeRange> getPlayed();
 
+  void setAllowBackgroundPlayback(bool allowBackgroundPlayback);
+
+  bool allowBackgroundPlayback();
+
   String? getError();
 
   void stop();
@@ -219,6 +223,9 @@ abstract class THEOplayerNativeAPI {
   void onLifecycleResume();
   
   void onLifecyclePause();
+
+  void configureSurface(int surfaceId, int width, int height);
+
 }