+++ flutter_theoplayer_sdk/CHANGELOG.md
@@ -1,6 +1,18 @@
+## 7.3.0
+
+* Updated THEOplayer to 7.3.0.
+* Added `AndroidConfig.viewComposition` to support Texture-based rendering.
+* Added `AndroidViewComposition.SURFACE_TEXTURE` to use SurfaceTexture on native Android (instead of PlatformViews).
+* Added `AndroidViewComposition.SURFACE_PRODUCER` to use SurfaceProducer on native Android (instead of PlatformViews).
+	- Only works from [Flutter 3.22.0](https://docs.flutter.dev/release/breaking-changes/android-surface-plugins#timeline)
+	- Supports [Texture-based rendering with Impeller](https://docs.flutter.dev/release/breaking-changes/android-surface-plugins#summary) rendering engine
+* Added Support for continuing playback when transitioning the app into background with `THEOplayer.setAllowBackgroundPlayback(boolean)`.
+* Added Fullscreen support with `THEOplayer.setPresentationMode(PresentationMode)`.
+* Deprecated `AndroidConfig.useHybridComposition` in favor of `AndroidConfig.viewComposition` (use `AndroidViewComposition.HYBRID_COMPOSITION` for the previous behaviour).
+
 ## 7.0.0
 
-* Update THEOplayer to 7.0.0.
+* Updated THEOplayer to 7.0.0.
 
 ### Versioning changes
 We are updating the version numbering for THEOplayer Flutter SDK to more closely match the underlying native THEOplayer SDKs.
@@ -14,8 +26,8 @@ Here are the notable changes:
 
 ## 1.0.3
 
-* Update THEOplayer to 6.10.1.
-* Add and couple application lifecycle listener.
+* Updated THEOplayer to 6.10.1.
+* Added and coupled application lifecycle listener.
 
 ## 1.0.2